# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: string;
  user_id: string;
  provider: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversation_count?: number | null;
  /** @nullable */
  last_import_at?: string | null;
  created_at: string;
}

export interface ProfileInput {
  provider: string;
  name: string;
  description?: string;
}

export interface ProfilePatch {
  name?: string;
  description?: string;
}

export interface ProfileStats {
  conversation_count: number;
  message_count: number;
  /** @nullable */
  last_import_at?: string | null;
  /** @nullable */
  total_size_bytes?: number | null;
}

export interface ImportRequest {
  profileId: string;
  provider: string;
}

export interface ImportReport {
  run_id: string;
  new_count: number;
  updated_count: number;
  skipped_count: number;
  failed_count: number;
  status: string;
  errors?: string[];
  /** @nullable */
  duration_ms?: number | null;
}

export interface ImportRun {
  id: string;
  profile_id: string;
  provider: string;
  status: string;
  /** @nullable */
  new_count?: number | null;
  /** @nullable */
  updated_count?: number | null;
  /** @nullable */
  skipped_count?: number | null;
  /** @nullable */
  failed_count?: number | null;
  created_at: string;
  /** @nullable */
  completed_at?: string | null;
}

export interface Conversation {
  id: string;
  profile_id: string;
  provider: string;
  display_title: string;
  /** @nullable */
  external_id?: string | null;
  /** @nullable */
  message_count?: number | null;
  /** @nullable */
  storage_path?: string | null;
  created_at: string;
  updated_at: string;
}

export interface ConversationDetail {
  id: string;
  profile_id: string;
  provider: string;
  display_title: string;
  /** @nullable */
  external_id?: string | null;
  /** @nullable */
  message_count?: number | null;
  /** @nullable */
  storage_path?: string | null;
  /** @nullable */
  markdown_content?: string | null;
  created_at: string;
  updated_at: string;
}

export interface ConversationPatch {
  display_title?: string;
}

export interface ConversationList {
  conversations: Conversation[];
  total: number;
  page: number;
  limit: number;
}

export interface DownloadUrl {
  url: string;
  /** @nullable */
  expires_at?: string | null;
}

export interface ProviderSummary {
  provider: string;
  profile_count: number;
  conversation_count: number;
  /** @nullable */
  last_import_at?: string | null;
}

export interface LibrarySummary {
  providers: ProviderSummary[];
  total_conversations?: number;
  total_profiles?: number;
}

export type ListProfilesParams = {
provider?: string;
};

export type ListImportRunsParams = {
profileId: string;
};

export type ListConversationsParams = {
profileId: string;
search?: string;
page?: number;
limit?: number;
};