# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationDetail {
  id: string;
  profile_id: string;
  provider: string;
  display_title: string;
  /** @nullable */
  external_id?: string | null;
  /** @nullable */
  message_count?: number | null;
  /** @nullable */
  storage_path?: string | null;
  /** @nullable */
  markdown_content?: string | null;
  created_at: string;
  updated_at: string;
}