# lib/api-zod/src/generated/types/conversationList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from './conversation';

export interface ConversationList {
  conversations: Conversation[];
  total: number;
  page: number;
  limit: number;
}