# lib/api-zod/src/generated/types/conversationPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationPatch {
  display_title?: string;
}