# lib/api-zod/src/generated/types/importReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */

export interface ImportReport {
  run_id: string;
  new_count: number;
  updated_count: number;
  skipped_count: number;
  failed_count: number;
  status: string;
  errors?: string[];
  /** @nullable */
  duration_ms?: number | null;
}