# lib/api-zod/src/generated/types/librarySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZenBrain API - AI Memory Vault
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderSummary } from './providerSummary';

export interface LibrarySummary {
  providers: ProviderSummary[];
  total_conversations?: number;
  total_profiles?: number;
}